backend/requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
openai
python-dotenv
